consolidate and simplify tests
