+++ test/view_component/view_component_test.rb
@@ -338,6 +338,54 @@ class ViewComponentTest < ViewComponent::TestCase
     assert_equal exception.message, "Validation failed: Content can't be blank"
   end
 
+  def test_compiles_unreferenced_component
+    assert UnreferencedComponent.compiled?
+  end
+
+  def test_does_not_compile_components_without_initializers
+    refute MissingInitializerComponent.compiled?
+  end
+
+  def test_raises_error_when_initializer_is_not_defined
+    exception = assert_raises ViewComponent::TemplateError do
+      render_inline(MissingInitializerComponent.new)
+    end
+
+    assert_includes exception.message, "must implement #initialize"
+  end
+
+  def test_raises_error_when_sidecar_template_is_missing
+    exception = assert_raises ViewComponent::TemplateError do
+      render_inline(MissingTemplateComponent.new)
+    end
+
+    assert_includes exception.message, "Could not find a template file for MissingTemplateComponent"
+  end
+
+  def test_raises_error_when_more_than_one_sidecar_template_is_present
+    error = assert_raises ViewComponent::TemplateError do
+      render_inline(TooManySidecarFilesComponent.new)
+    end
+
+    assert_includes error.message, "More than one template found for TooManySidecarFilesComponent."
+  end
+
+  def test_raises_error_when_more_than_one_sidecar_template_for_a_variant_is_present
+    error = assert_raises ViewComponent::TemplateError do
+      render_inline(TooManySidecarFilesForVariantComponent.new)
+    end
+
+    assert_includes error.message, "More than one template found for variants 'test' and 'testing' in TooManySidecarFilesForVariantComponent"
+  end
+
+  def test_backtrace_returns_correct_file_and_line_number
+    error = assert_raises NameError do
+      render_inline(ExceptionInTemplateComponent.new)
+    end
+
+    assert_match %r[app/components/exception_in_template_component\.html\.erb:2], error.backtrace[0]
+  end
+
   private
 
   def modify_file(file, content)
